Limit www canister to just ads_ledger canister (#11)
Candid UI by default connects to any canister specified by the `id` GET
variable. Since this instance of the canister is meant to be
single-purpose, it would make sense to access that canister by default
and remove the option to switch.

Along with that, some changes to replace "Candid UI" with "Ethical
Targeting", just to emphasize what exactly is being accessed here.